+++ src/components/AckModalLayout.component.ts
@@ -53,14 +53,14 @@ import { string } from "./templates/ack-modal-layout.pug"
   }
 
   ngOnInit(){
-    setTimeout(()=>{
+    Promise.resolve().then(()=>{
       if( this.isModelMode || (this.isModelMode==null && this.showModelChange.observers.length) ){
         this.isModelMode = true
       }
 
       this.backgroundColor = this.backgroundColor || 'rgba(255,255,255,0.95)'
       //this.backgroundColorChange.emit(this.backgroundColor)
-    }, 0)
+    })
   }
 
   fireClose(){