+++ src/directives/FxOn.directive.ts
@@ -30,7 +30,7 @@ import {
   constructor(public element:ElementRef){}
 
   ngOnInit(){
-    setTimeout(()=>this.update(), 0)
+    Promise.resolve().then(()=>this.update())
   }
 
   update(){